add AddFilter.js
