+++ src/Components/Filters/FilterBox.js
@@ -21,6 +21,7 @@ import Grid from "@material-ui/core/Grid";
 
 import CompareSubSamples from "./CompareSubsamples";
 import DataSeriesFilter from "./DataSeriesFilter";
+import AddFilter from "./AddFilter";
 
 export default function FilterBox(props) {
   const History = useHistory();
@@ -211,55 +212,6 @@ export default function FilterBox(props) {
           colourStyles={colourOptions}
         />
       );
-      // return (
-      //   <>
-      //     {/* labels filter */}
-      //     <p>{filterSelectorName}</p>
-      //     <Select
-      //       defaultValue={{ label: filters[index].selectedCategory }}
-      //       // isClearable
-      //       //seems not in use
-      //       formatGroupLabel={formatGroupLabel}
-      //       components={{ Control: ControlComponent }}
-      //       // isSearchable
-      //       onChange={e => {
-      //         setUpdateUrlFlag(!updateUrlFlag);
-      //         let optionFlags = {};
-      //         graphLabels[
-      //           `${FilterBoxOptions.default[e.label].value.type}`
-      //         ].labels.forEach(option => {
-      //           optionFlags = {
-      //             ...optionFlags,
-      //             [option]: false
-      //           };
-      //         });
-      //         setFilters({
-      //           ...filters,
-      //           [index]: {
-      //             ...filters[index],
-      //             selectedCategory: e.label, //option
-      //             selectedTableColumnName:
-      //               FilterBoxOptions.default[e.label].value.type,
-
-      //             selectedTable: FilterBoxOptions.default[e.label].value.query,
-      //             selectedOption: undefined,
-      //             selectableOptions: { ...optionFlags }
-      //           }
-      //         });
-      //       }}
-      //       name="color"
-      //       styles={colourStyles}
-      //       options={xVar(
-      //         FilterBoxOptions.superCategories,
-      //         Object.keys(filters)
-      //           .map(filterId => {
-      //             return filters[filterId].selectedCategory;
-      //           })
-      //           .filter(selectedCategory => selectedCategory.length > 0)
-      //       )}
-      //     />
-      //   </>
-      // );
       //render Compare SubSamples if on BarChart
     } else if (filterSelectorName === "Compare SubSamples" && open === "bar") {
       return (
@@ -283,66 +235,21 @@ export default function FilterBox(props) {
       return <></>;
     } else if (filterSelectorName === "Data Filter") {
       return (
-        <div>
-          <form>
-            {/* labels filter */}
-            <p>{filterSelectorName}</p>
-            <Select
-              defaultValue={{ label: filters[index].selectedCategory }}
-              // isClearable
-              //seems not in use
-              formatGroupLabel={formatGroupLabel}
-              components={{ Control: ControlComponent }}
-              // isSearchable
-              onChange={e => {
-                setUpdateUrlFlag(!updateUrlFlag);
-                let optionFlags = {};
-                graphLabels[
-                  `${FilterBoxOptions.default[e.label].value.type}`
-                ].labels.forEach(option => {
-                  optionFlags = {
-                    ...optionFlags,
-                    [option]: false
-                  };
-                });
-                setFilters({
-                  ...filters,
-                  [index]: {
-                    ...filters[index],
-                    selectedCategory: e.label, //option
-                    selectedTableColumnName:
-                      FilterBoxOptions.default[e.label].value.type,
-
-                    selectedTable:
-                      FilterBoxOptions.default[e.label].value.query,
-                    selectedOption: undefined,
-                    selectableOptions: { ...optionFlags }
-                  }
-                });
-              }}
-              name="color"
-              styles={colourStyles}
-              options={dataFilterVar(
-                FilterBoxOptions.superCategories,
-                Object.keys(filters)
-                  .map(filterId => {
-                    return filters[filterId].selectedCategory;
-                  })
-                  .filter(selectedCategory => selectedCategory.length > 0)
-              )}
-            />
-            {/* additional options below 'Data Series' and 'Add Filter' */}
-            <RenderCheckContainer
-              i={index}
-              filters={filters}
-              graphLabels={graphLabels}
-              props={props}
-              CategoryOptions={CategoryOptions}
-              CheckboxContainer={CheckboxContainer}
-              setFilters={setFilters}
-            />
-          </form>
-        </div>
+        <AddFilter
+          filterSelectorName={filterSelectorName}
+          filters={filters}
+          setFilters={setFilters}
+          formatGroupLabel={formatGroupLabel}
+          ControlComponent={ControlComponent}
+          index={index}
+          formatGroupLabel={formatGroupLabel}
+          setUpdateUrlFlag={setUpdateUrlFlag}
+          FilterBoxOptions={FilterBoxOptions}
+          updateUrlFlag={updateUrlFlag}
+          dataFilterVar={dataFilterVar}
+          colourStyles={colourOptions}
+          CategoryOptions={CategoryOptions}
+        />
       );
     }
   };
